audio_modem/requirements.txt
cnumpy
scipy
matplotlib  # (optional for visual debugging)
pyaudio     # (optional for live audio capture/playback)
